# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Assemble Yourself",
        "description": "Приложение для личностного роста",
        "contact": {
            "email": "v.a.kuznetsov-83@yandex.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses/{id}/content": {
            "get": {
                "tags": [
                    "content"
                ],
                "summary": "Получение контента курса",
                "description": "Получает курсы по ID курса",
                "operationId": "getContentByCourseId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID курса по которому выберается контент",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Отсутсвует содержимое курса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Отсутсвует содержимое курса"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "content"
                ],
                "summary": "Изменение содержания курса",
                "description": "Изменяет или создает контент курса с заданными параметрами",
                "operationId": "createContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID по которому выберается курс",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "ContentArray",
                    "description": "Перечень свойств контента",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "description": "Содержание курса",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Переданы не все обязательные поля",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Поле cousre_id обязательно для заполнения"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Содержание курса может создавать только авторизованый пользователь"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Пользователь не автор курса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Содержание курса может создавать только автор курса"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Курса не существует",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Такого курса не существует"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ContentModel": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "course_id": {
                        "description": "ID курса",
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "content": {
                        "description": "Содержание курса",
                        "type": "string",
                        "readOnly": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Content"
                }
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer {Token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}